requirements.txt
brother_ql
bottle
jinja2
streamlit
flask
pillow